% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-cat.R
\name{barplot.freq_table}
\alias{barplot.freq_table}
\title{Frequency Table Bar Plot}
\usage{
\method{barplot}{freq_table}(height, ...)
}
\arguments{
\item{height}{an object of class \code{freq_table}}

\item{...}{further arguments to be passed to or from methods}
}
\description{
\code{barplot.freq_table} creates bar plot
for the frequency table created using \code{freq_table}
}
\examples{
k <- freq_table(mtcars$cyl)
barplot(k)
}

