% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom-dist.R
\name{binom_perc}
\alias{binom_perc}
\title{Binomial Distribution Percentile}
\usage{
binom_perc(n, p, tp, type = c("lower", "upper"))
}
\arguments{
\item{n}{number of bernoulli trials}

\item{p}{a single aggregated probability of multiple trials}

\item{tp}{the probability of success in a trial}

\item{type}{lower tail or upper tail}
}
\description{
Calculate and visualize quantiles out of given probability
}
\examples{
binom_perc(10, 0.5, 0.05)
binom_perc(10, 0.5, 0.05, "upper")
}
\seealso{
\code{\link{binom_plot}} \code{\link{binom_prob}}
\code{\link[stats]{Binomial}}
}

