% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom-dist.R
\name{binom_plot}
\alias{binom_plot}
\title{Visualize Binomial Distribution}
\usage{
binom_plot(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success on each trial}
}
\value{
a list containing the following components:

\item{avg}{mean of the binomial distribution}
\item{stdev}{standard deviation of the binomial distribution}
}
\description{
Visualize how changes in number of trials and the probability of
success affect the shape of the binomial distribution.
}
\examples{
binom_plot(10, 0.3)
}
\seealso{
\code{\link{binom_perc}} \code{\link{binom_prob}}
\code{\link[stats]{Binomial}}
}

