% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-summary.R
\name{boxplot.group_summary}
\alias{boxplot.group_summary}
\title{Group Summary Box Plot}
\usage{
\method{boxplot}{group_summary}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{group_summary}}
}
\description{
\code{boxplot.group_summary} creates boxplots of the continuous
variable for the different levels of the categorical variable.
}
\examples{
mt <- mtcars
mt$cyl <- as.factor(mt$cyl)
k <- group_summary(mt$cyl, mt$mpg)
boxplot(k)
}

