% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-dist.R
\name{chi_per}
\alias{chi_per}
\title{Chi Square Distribution Percentile}
\usage{
chi_per(probs = 0.95, df = 3, type = c("lower", "upper"))
}
\arguments{
\item{probs}{a probability value}

\item{df}{degrees of freedom}

\item{type}{lower tail or upper tail}
}
\value{
percentile for the \code{probs} based on \code{df} and \code{type}
}
\description{
Visualize the percentile from the value of the lower/upper
cumulative distribution function of the chi square distribution
}
\examples{
chi_per(0.165, 8, 'upper')
chi_per(0.22, 13, 'upper')
}
\seealso{
\code{\link{chi_plot}} \code{\link{chi_prob}}
\code{\link[stats]{Chisquare}}
}

