% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-dist.R
\name{chi_plot}
\alias{chi_plot}
\title{Visualize Chi Square Distribution}
\usage{
chi_plot(df = 3, normal = FALSE)
}
\arguments{
\item{df}{degrees of freedom}

\item{normal}{logical; if \code{TRUE}, normal curve with same \code{mean} and
\code{sd} as the chi square distribution is drawn}
}
\description{
Visualize how changes in degrees of freedom affect the shape of
the chi square distribution.
}
\examples{
chi_plot()
chi_plot(df = 5)
chi_plot(df = 5, normal = TRUE)
}
\seealso{
\code{\link{chi_per}} \code{\link{chi_prob}}
\code{\link[stats]{Chisquare}}
}

