% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-binomial.R
\name{dist_binom_plot}
\alias{dist_binom_plot}
\alias{dist_binom_prob}
\alias{dist_binom_perc}
\title{Visualize binomial distribution}
\usage{
dist_binom_plot(n, p)

dist_binom_prob(n, p, s, type = c("lower", "upper", "exact", "interval"))

dist_binom_perc(n, p, tp, type = c("lower", "upper"))
}
\arguments{
\item{n}{Number of trials.}

\item{p}{Aggregate probability.}

\item{s}{Number of success.}

\item{type}{Lower/upper/exact/interval.}

\item{tp}{Probability of success in a trial.}
}
\value{
A list containing the following components:

\item{avg}{Mean of the binomial distribution,}
\item{stdev}{Standard deviation of the binomial distribution.}
\item{prob}{Probability of success.}
}
\description{
Visualize how changes in number of trials and the probability of
success affect the shape of the binomial distribution. Compute & visualize
probability from a given quantile and quantiles out of given probability.
}
\examples{
# visualize binomial distribution
dist_binom_plot(10, 0.3)

# visualize probability from a given quantile
dist_binom_prob(10, 0.3, 4, type = 'exact')
dist_binom_prob(10, 0.3, 4, type = 'lower')
dist_binom_prob(10, 0.3, 4, type = 'upper')
dist_binom_prob(10, 0.3, c(4, 6), type = 'interval')

# visualize quantiles out of given probability
dist_binom_perc(10, 0.5, 0.05)
dist_binom_perc(10, 0.5, 0.05, "upper")

}
\seealso{
\code{\link[stats]{Binomial}}
}
