% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-auto-summary.R
\name{ds_auto_group_summary}
\alias{ds_auto_group_summary}
\title{Tabulation}
\usage{
ds_auto_group_summary(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Generate summary statistics for all continuous variables in data.
}
\examples{
# summary statistics of mpg & disp for each level of cyl & gear
ds_auto_group_summary(mtcarz, cyl, gear, mpg, disp)

}
