% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_css}
\alias{ds_css}
\title{Corrected Sum of Squares}
\usage{
ds_css(x, data = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{data}{a \code{data.frame} or \code{tibble}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Compute the corrected sum of squares
}
\examples{
ds_css(mtcars$mpg)
ds_css(mpg, mtcars)
}
