% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_extreme_obs}
\alias{ds_extreme_obs}
\title{Extreme observations}
\usage{
ds_extreme_obs(data, column)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{column}{Column in \code{data}.}
}
\description{
Returns the most extreme observations.
}
\examples{
ds_extreme_obs(mtcarz, mpg)

}
