% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_gmean}
\alias{ds_gmean}
\title{Geometric Mean}
\usage{
ds_gmean(data, x = NULL)
}
\arguments{
\item{data}{A numeric vector or \code{data.frame}.}

\item{x}{Column in \code{data}.}
}
\description{
Computes the geometric mean
}
\examples{

# vector
ds_gmean(mtcars$mpg)

# data.frame
ds_gmean(mtcars, mpg)

}
\seealso{
\code{\link{ds_hmean}} \code{\link[base]{mean}}
}
