% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_hmean}
\alias{ds_hmean}
\title{Harmonic Mean}
\usage{
ds_hmean(data, x = NULL)
}
\arguments{
\item{data}{A numeric vector or \code{data.frame}.}

\item{x}{Column in \code{data}.}
}
\description{
Computes the harmonic mean
}
\examples{
# vector
ds_hmean(mtcars$mpg)

# data.frame
ds_hmean(mtcars, mpg)

}
\seealso{
\code{\link{ds_gmean}} \code{\link[base]{mean}}
}
