% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_kurtosis}
\alias{ds_kurtosis}
\title{Kurtosis}
\usage{
ds_kurtosis(x, data = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{data}{a \code{data.frame} or \code{tibble}}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Compute the kurtosis of a probability distribution.
}
\examples{
ds_kurtosis(mtcars$mpg)
ds_kurtosis(mpg, mtcars)
}
\references{
Sheskin, D.J. (2000) Handbook of Parametric and Nonparametric Statistical Procedures, Second Edition. Boca Raton, Florida: Chapman & Hall/CRC.
}
\seealso{
\code{ds_skewness}
}
