% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-launch-shiny-app.R
\name{ds_launch_shiny_app}
\alias{ds_launch_shiny_app}
\title{Launch Shiny App}
\usage{
ds_launch_shiny_app()
}
\description{
Launches shiny app
}
\section{Deprecated Function}{

\code{launch_descriptr()} has been deprecated. Instead
use \code{ds_launch_shiny_app()}.
}

\examples{
\dontrun{
ds_launch_shiny_app()
}
}
