% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{ds_mdev}
\alias{ds_mdev}
\alias{stat_mdev}
\title{Mean Absolute Deviation}
\usage{
ds_mdev(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
Mean absolute deviation of \code{x}
}
\description{
Compute the mean absolute deviation about the mean
}
\details{
The \code{stat_mdev} function computes the mean absolute deviation
about the mean. It is different from \code{mad} in \code{stats} package as
the statistic used to compute the deviations is not \code{median} but
\code{mean}. Any NA values are stripped from \code{x} before computation
takes place
}
\section{Deprecated Function}{

\code{stat_mdev()} has been deprecated. Instead use \code{ds_mdev()}.
}

\examples{
ds_mdev(mtcars$mpg)
}
\seealso{
\code{\link[stats]{mad}}
}
