% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_mode}
\alias{ds_mode}
\alias{stat_mode}
\title{Mode}
\usage{
ds_mode(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose mode is to be computed}
}
\value{
Mode of \code{x}
}
\description{
Compute the sample mode
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\section{Deprecated Function}{

`stat_mode()` has been deprecated. Instead use `ds_mode()`.
}

\examples{
ds_mode(mtcars$mpg)
ds_mode(mtcars$cyl)
}
\seealso{
\code{\link[base]{mean}} \code{\link[stats]{median}}
}
