% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-summary-stats.R
\name{ds_multi_summary_stats}
\alias{ds_multi_summary_stats}
\title{Statistics for multiple continuous variables}
\usage{
ds_multi_summary_stats(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Range of descriptive statistics for multiple continuous variables.
}
\examples{
ds_multi_summary_stats(mtcarz)
ds_multi_summary_stats(mtcarz, mpg, disp)

}
\seealso{
\code{\link[base]{summary}} \code{\link{ds_freq_cont}}
\code{\link{ds_freq_table}} \code{\link{ds_cross_table}}
}
