% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-mult-table.R
\name{ds_oway_tables}
\alias{ds_oway_tables}
\alias{oway_tables}
\alias{ds_tway_tables}
\alias{tway_tables}
\title{Multiple One & Two Way Tables}
\usage{
ds_oway_tables(data)

ds_tway_tables(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
\code{ds_oway_tables} creates multiple one way tables by creating
a frequency table for each categorical variable in a data frame.
\code{ds_tway_tables} creates multiple two way tables by creating a cross
table for each unique pair of categorical variables in a data frame.
}
\details{
\code{ds_oway_tables} is a extension of the \code{ds_freq_table}
function. It creates a frequency table for each categorical variable in the
dataframe. \code{ds_tway_tables} is a extension of the \code{ds_cross_table}
function. It creates a two way table for each unique pair of categorical
variables in the dataframe.
}
\section{Deprecated Functions}{

\code{oway_tables()} and \code{tway_tables()} have been deprecated. Instead
use \code{ds_oway_tables()} and \code{ds_tway_tables()}.
}

\examples{
mt <- mtcars
mt[, c(2, 8, 9)] <- lapply(mt[, c(2, 8, 9)], factor)
# multiple one way tables
ds_oway_tables(mt)

# multiple two way tables
ds_tway_tables(mt)
}
\seealso{
\code{link{ds_freq_table}} \code{link{ds_cross_table}}
}
