% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_bar}
\alias{ds_plot_bar}
\title{Generate bar plots}
\usage{
ds_plot_bar(data, ..., fill = "blue")
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}

\item{fill}{Color of the bars.}
}
\description{
Creates bar plots if the data has categorical variables.
}
\examples{
ds_plot_bar(mtcarz)
ds_plot_bar(mtcarz, cyl)
ds_plot_bar(mtcarz, cyl, gear)

}
