% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_bar_stacked}
\alias{ds_plot_bar_stacked}
\title{Generate stacked bar plots}
\usage{
ds_plot_bar_stacked(data, ..., print_plot = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Creates stacked bar plots if the data has categorical variables.
}
\examples{
# subset data
mt <- dplyr::select(mtcarz, cyl, gear, am)

# stacked bar plot
ds_plot_bar_stacked(mtcarz, cyl, gear)

# plot all variables
ds_plot_bar_stacked(mt)

}
