% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_box_single}
\alias{ds_plot_box_single}
\title{Generate box plots}
\usage{
ds_plot_box_single(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Creates box plots if the data has continuous variables.
}
\examples{
ds_plot_box_single(mtcarz)
ds_plot_box_single(mtcarz, mpg)
ds_plot_box_single(mtcarz, mpg, disp, hp)

}
