% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_density}
\alias{ds_plot_density}
\title{Generate density plots}
\usage{
ds_plot_density(data, ..., color = "blue")
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}

\item{color}{Color of the plot.}
}
\description{
Creates density plots if the data has continuous variables.
}
\examples{
ds_plot_density(mtcarz)
ds_plot_density(mtcarz, mpg)
ds_plot_density(mtcarz, mpg, disp, hp)

}
