% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_histogram}
\alias{ds_plot_histogram}
\title{Generate histograms}
\usage{
ds_plot_histogram(data, ..., bins = 5, fill = "blue", print_plot = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}

\item{bins}{Number of bins in the histogram.}

\item{fill}{Color of the histogram.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Creates histograms if the data has continuous variables.
}
\examples{
ds_plot_histogram(mtcarz)
ds_plot_histogram(mtcarz, mpg)
ds_plot_histogram(mtcarz, mpg, disp, hp)

}
