% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_scatter}
\alias{ds_plot_scatter}
\title{Generate scatter plots}
\usage{
ds_plot_scatter(data, ..., print_plot = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Creates scatter plots if the data has continuous variables.
}
\examples{
ds_plot_scatter(mtcarz)
ds_plot_scatter(mtcarz, mpg, disp)

}
