% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_skewness}
\alias{ds_skewness}
\title{Skewness}
\usage{
ds_skewness(data, x = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A numeric vector or \code{data.frame}.}

\item{x}{Column in \code{data}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Compute the skewness of a probability distribution.
}
\examples{

# vector
ds_skewness(mtcars$mpg)

# data.frame
ds_skewness(mpg, mtcars)

}
\references{
Sheskin, D.J. (2000) Handbook of Parametric and Nonparametric Statistical Procedures, Second Edition. Boca Raton, Florida: Chapman & Hall/CRC.
}
\seealso{
\code{kurtosis}
}
