% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-utils.R
\name{ds_std_error}
\alias{ds_std_error}
\title{Standard error of mean}
\usage{
ds_std_error(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Returns the standard error of mean.
}
\examples{
ds_std_error(mtcars$mpg)

}
