% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-summary-stats.R
\name{ds_summary_stats}
\alias{ds_summary_stats}
\alias{summary_stats}
\title{Descriptive statistics}
\usage{
ds_summary_stats(data, variable)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{variable}{Column in \code{data}.}
}
\description{
Range of descriptive statistics for continuous data.
}
\section{Deprecated function}{

\code{summary_stats()} has been deprecated. Instead use
\code{ds_summary_stats()}.
}

\examples{
ds_summary_stats(mtcarz, mpg)

}
\seealso{
\code{\link[base]{summary}} \code{\link{ds_freq_cont}}
\code{\link{ds_freq_table}} \code{\link{ds_cross_table}}
}
