% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-multistats.R
\name{ds_tidy_stats}
\alias{ds_tidy_stats}
\alias{ds_multi_stats}
\title{Tidy descriptive statistics}
\usage{
ds_tidy_stats(data, ...)
}
\arguments{
\item{data}{A \code{tibble} or a \code{data.frame}.}

\item{...}{Columns in \code{x}.}
}
\value{
A tibble.
}
\description{
Descriptive statistics for multiple variables.
}
\section{Deprecated Functions}{

\code{ds_multi_stats()} have been deprecated. Instead use \code{ds_tidy_stats()}.
}

\examples{
ds_tidy_stats(mtcarz)
ds_tidy_stats(mtcarz, mpg, disp, hp)

}
