% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f-dist.R
\name{f_per}
\alias{f_per}
\title{Visualize F Distribution Percentile}
\usage{
f_per(probs = 0.95, num_df = 3, den_df = 30, type = c("lower", "upper"))
}
\arguments{
\item{probs}{a probability value}

\item{num_df}{degrees of freedom associated with the numerator of f statistic}

\item{den_df}{degrees of freedom associated with the denominator of f statistic}

\item{type}{lower tail or upper tail}
}
\value{
percentile for the \code{probs} based on \code{num_df}, \code{den_df}
and \code{type}
}
\description{
Calculate and visualize quantiles out of given probability
}
\examples{
f_per(0.95, 3, 30, 'lower')
f_per(0.125, 9, 35, 'upper')
}
\seealso{
\code{\link{f_plot}} \code{\link{f_prob}} \code{\link[stats]{FDist}}
}

