% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f-dist.R
\name{f_plot}
\alias{f_plot}
\title{Visualize F Distribution}
\usage{
f_plot(num_df = 4, den_df = 30, normal = FALSE)
}
\arguments{
\item{num_df}{degrees of freedom associated with the numerator of f statistic}

\item{den_df}{degrees of freedom associated with the denominator of f statistic}

\item{normal}{logical; if \code{TRUE}, normal curve with same \code{mean} and
\code{sd} as the F distribution is drawn.}
}
\description{
Visualize how changes in degrees of freedom affect the
shape of the F distribution.
}
\examples{
f_plot()
f_plot(6, 10, normal = TRUE)
}
\seealso{
\code{\link{f_per}} \code{\link{f_prob}}
\code{\link[stats]{FDist}}
}

