% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f-dist.R
\name{f_prob}
\alias{f_prob}
\title{Visualize F Distribution Probabilities}
\usage{
f_prob(perc, num_df, den_df, type = c("lower", "upper"))
}
\arguments{
\item{perc}{a quantile value}

\item{num_df}{degrees of freedom associated with the numerator of f statistic}

\item{den_df}{degrees of freedom associated with the denominator of f statistic}

\item{type}{lower tail or upper tail}
}
\value{
probability value for \code{perc} based on \code{num_df}, \code{den_df}
and \code{type}
}
\description{
Calculate and visualize probability from a given quantile
}
\examples{
f_prob(2.35, 5, 32)
f_prob(1.5222, 9, 35, type = "upper")
}
\seealso{
\code{\link{f_plot}} \code{\link{f_per}} \code{\link[stats]{FDist}}
}

