% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-cont.R
\name{freq_cont}
\alias{freq_cont}
\alias{freq_cont.default}
\title{Frequency Distribution of Continuous Data}
\usage{
freq_cont(data, bins = 5)

\method{freq_cont}{default}(data, bins = 5)
}
\arguments{
\item{data}{numeric vector}

\item{bins}{number of intervals into which the data must be split}
}
\value{
\code{freq_cont} returns an object of class \code{"freq_cont"}.
An object of class \code{"freq_cont"} is a list containing the
following components

\item{breaks}{lower/upper boundaries of intervals}
\item{frequency}{frequecy of the intervals}
\item{cumulative}{cumulative frequency}
\item{percent}{frequency as percent}
\item{cum_percent}{cumulative frequency as percent}
\item{bins}{bins}
\item{data}{data}
\item{varname}{name of the data}
}
\description{
\code{freq_cont} returns the frequency distribution of
continuous by splitting the data into equidistant intervals created based on
the number of bins specified.
}
\examples{
freq_cont(mtcars$mpg, 4)
}
\seealso{
\code{link{freq_table}} \code{link{cross_table}}
}

