% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-cat.R
\name{freq_table}
\alias{freq_table}
\alias{freq_table.default}
\title{Frequency Table: Categorical Data}
\usage{
freq_table(data)

\method{freq_table}{default}(data)
}
\arguments{
\item{data}{numeric or factor vector}
}
\value{
\code{freq_table} returns an object of class \code{"freq_table"}.
An object of class \code{"freq_table"} is a list containing the
following components

\item{ftable}{frequency table}
\item{varname}{name of the data}
}
\description{
\code{freq_table} creates frequency table for factor data and
returns the frequency, cumulative frequency, frequency percent and cumulative
frequency percent.
}
\examples{
freq_table(mtcars$cyl)
freq_table(as.factor(mtcars$cyl))
}
\seealso{
\code{link{freq_cont}} \code{link{cross_table}}
}

