% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-summary.R
\name{group_summary}
\alias{group_summary}
\alias{group_summary.default}
\title{Descriptive Statistics By Group}
\usage{
group_summary(x, y)

\method{group_summary}{default}(x, y)
}
\arguments{
\item{x}{a factor variable}

\item{y}{a continuous variable}
}
\value{
\code{group_summary} returns an object of class \code{"group_summary"}.
An object of class \code{"group_summary"} is a list containing the
following components:

\item{stats}{a data frame containing descriptive statistics for the different
levels of the factor variable}
\item{plotdata}{data for boxplot method}
\item{xvar}{name of the categorical variable}
\item{yvar}{name of the continuous variable}
}
\description{
\code{group_summary} returns descriptive statistics of a
continuous variable for the different levels of a categorical variable.
}
\examples{
mt <- mtcars
mt$cyl <- as.factor(mt$cyl)
group_summary(mt$cyl, mt$mpg)
}
\seealso{
\code{link{summary_stats}}
}

