% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq-cont.R
\name{hist.freq_cont}
\alias{hist.freq_cont}
\title{Frequency Table Histogram}
\usage{
\method{hist}{freq_cont}(x, col = "blue", ...)
}
\arguments{
\item{x}{an object of class \code{freq_cont}}

\item{col}{color of the bars}

\item{...}{further arguments to be passed to or from methods}
}
\description{
\code{hist.freq_cont} creates histogram
for the frequency table created using \code{freq_cont}
}
\examples{
k <- freq_cont(mtcars$mpg, 4)
plot(k)
}

