% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{hmean}
\alias{hmean}
\title{Harmonic Mean}
\usage{
hmean(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose harmonic mean is to be
computed}
}
\value{
Returns the harmonic mean of x
}
\description{
Compute the harmonic mean
}
\examples{
hmean(mtcars$mpg)
}
\seealso{
\code{\link{gmean}}
}

