% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose kurtosis is to be computed}
}
\value{
Kurtosis of \code{x}
}
\description{
Compute the kurtosis of a probability distribution.
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
kurtosis(mtcars$mpg)
}
\references{
Sheskin, D.J. (2000) Handbook of Parametric and Nonparametric Statistical Procedures, Second Edition. Boca Raton, Florida: Chapman & Hall/CRC.
}
\seealso{
\code{skewness}
}

