% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-table.R
\name{mosaicplot.cross_table}
\alias{mosaicplot.cross_table}
\title{Mosaic Plot Cross Table}
\usage{
\method{mosaicplot}{cross_table}(x, ...)
}
\arguments{
\item{x}{An object of class cross_table}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
\code{mosaicplot.cross_table} creates mosaic plot
for the two way tables created using \code{cross_table}
}
\details{
Mosaic plot method added to cross_table
}
\examples{
k <- cross_table(mtcars$cyl, mtcars$am)
mosaicplot(k)

}

