% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-data-mtcarz.R
\docType{data}
\name{mtcarz}
\alias{mtcarz}
\title{mtcarz}
\format{An object of class \code{data.frame} with 32 rows and 11 columns.}
\usage{
mtcarz
}
\description{
Copy of mtcars data set with modified variable types
}
\keyword{datasets}
