% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistats.R
\name{multistats}
\alias{multistats}
\title{Multi Variable Statistics}
\usage{
multistats(x, ...)
}
\arguments{
\item{x}{a tibble/data.frame}

\item{...}{columns in \code{x}}
}
\value{
a tibble
}
\description{
Summary/descriptive statistics for multiple variables
}
\examples{
multistats(mtcars, mpg, disp, hp)
}
