% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-dist.R
\name{norm_per}
\alias{norm_per}
\title{Visualize Nomral Distribution Percentile}
\usage{
norm_per(probs = 0.95, mean = 0, sd = 1, type = c("lower", "upper",
  "both"))
}
\arguments{
\item{probs}{a probability value}

\item{mean}{mean of the normal distribution}

\item{sd}{standard deviation of the normal distribution}

\item{type}{lower tail, upper tail or both}
}
\value{
percentile for the \code{probs} based on \code{mean}, \code{sd} and
\code{type}
}
\description{
Calculate and visualize quantiles out of given probability
}
\examples{
norm_per(0.95, mean = 2, sd = 1.36)
norm_per(0.3, mean = 2, sd = 1.36, type = 'upper')
norm_per(0.95, mean = 2, sd = 1.36, type = 'both')
}
\seealso{
\code{\link{norm_plot}} \code{\link{norm_prob}} \code{\link[stats]{Normal}}
}

