% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-dist.R
\name{norm_plot}
\alias{norm_plot}
\title{Visualize Normal Distribution}
\usage{
norm_plot(mean = 0, sd = 1)
}
\arguments{
\item{mean}{mean of the normal distribution}

\item{sd}{standard deviation of the normal distribution}
}
\description{
Visualize how changes in mean and standard deviation affect the
shape of the normal distribution.
}
\examples{
norm_plot()
norm_plot(mean = 2, sd = 0.6)
}
\seealso{
\code{\link{norm_per}} \code{\link{norm_prob}}
\code{\link[stats]{Normal}}
}

