% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-dist.R
\name{norm_prob}
\alias{norm_prob}
\title{Visualize Normal Distribution Probabilities}
\usage{
norm_prob(perc, mean = 0, sd = 1, type = c("lower", "upper", "both"))
}
\arguments{
\item{perc}{a quantile value}

\item{mean}{mean of the normal distribution}

\item{sd}{standard deviation of the normal distribution}

\item{type}{lower tail, upper tail, or both}
}
\value{
probability value for \code{perc} based on \code{mean}, \code{sd} and
\code{type}
}
\description{
Calculate and visualize probability from a given quantile
}
\examples{
norm_prob(3.78, mean = 2, sd = 1.36)
norm_prob(3.43, mean = 2, sd = 1.36, type = 'upper')
norm_prob(c(-1.74, 1.83), type = 'both')
}
\seealso{
\code{\link{norm_plot}} \code{\link{norm_per}} \code{\link[stats]{Normal}}
}

