% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult-table.R
\name{oway_tables}
\alias{oway_tables}
\title{Multiple One Way Tables}
\usage{
oway_tables(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
\code{oway_tables} creates multiple one way tables by creating
a frequency table for each categorical variable in a data frame.
}
\details{
\code{oway_tables} is a extension of the \code{freq_table} 
function. It creates a frequency table for each categorical variable in the 
dataframe.
}
\examples{
mt <- mtcars
mt[, c(2, 8:11)] <- lapply(mt[, c(2, 8:11)], factor)
oway_tables(mt)
}
\seealso{
\code{link{tway_tables}} \code{link{freq_table}} 
\code{link{cross_table}}
}

