% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-table.R
\name{plot.cross_table}
\alias{plot.cross_table}
\title{Cross Table Bar Plot}
\usage{
\method{plot}{cross_table}(x, beside = FALSE, proportional = FALSE, ...)
}
\arguments{
\item{x}{An object of class cross_table}

\item{beside}{a logical value. If FALSE, the columns of height are portrayed
as stacked bars, and if TRUE the columns are portrayed as juxtaposed bars.}

\item{proportional}{a logical value. If TRUE, the height of the bars is
proportional}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
\code{barplot.cross_table} creates stacked and grouped bar plots
for the two way tables created using \code{cross_table}
}
\details{
Bar plot method added to cross_table
}
\examples{
k <- cross_table(mtcars$cyl, mtcars$am)
plot(k)
plot(k, beside = TRUE)
plot(k, proportional = TRUE)

}

