% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screener.R
\name{plot.screener}
\alias{plot.screener}
\title{Visualize Missing Values}
\usage{
\method{plot}{screener}(x, ...)
}
\arguments{
\item{x}{an object of class \code{screener}}

\item{...}{further arguments to be passed to or from methods}
}
\description{
\code{plot.screener} creates bar plots to visualize % of missing
observations for each variable in a data frame
}
\examples{
mt <- mtcars
mt[, c(2, 8:11)] <- lapply(mt[, c(2, 8:11)], factor)
mt[sample(1:n, 12), sample(1:cl, 6)] <- NA
k <- screener(mt)
plot(k)
}

