% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose skewness is to be computed}
}
\value{
Skewness of \code{x}
}
\description{
Compute the skewness of a probability distribution.
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
skewness(mtcars$mpg)
}
\references{
Sheskin, D.J. (2000) Handbook of Parametric and Nonparametric Statistical Procedures, Second Edition. Boca Raton, Florida: Chapman & Hall/CRC.
}
\seealso{
\code{kurtosis}
}
