% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{stat_css}
\alias{stat_css}
\title{Corrected Sum of Squares}
\usage{
stat_css(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose mode is to be computed}
}
\value{
Corrected sum of squares of \code{x}
}
\description{
Compute the corrected sum of squares
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
stat_css(mtcars$mpg)
}
\references{
\href{http://www.itl.nist.gov/div898/handbook/prc/section4/prc421.htm}{NIST/SEMATECH e-Handbook of Statistical Methods}
}
