% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{stat_cvar}
\alias{stat_cvar}
\title{Coefficient of Variation}
\usage{
stat_cvar(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose mode is to be computed}
}
\description{
Compute the coefficient of variation
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
stat_cvar(mtcars$mpg)
}
\seealso{
\code{\link[stats]{var}} \code{\link[stats]{sd}}
}

