% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{stat_mode}
\alias{stat_mode}
\title{Mode}
\usage{
stat_mode(x)
}
\arguments{
\item{x}{a numeric vector containing the values whose mode is to be computed}
}
\value{
Mode of \code{x}
}
\description{
Compute the sample mode
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
stat_mode(mtcars$mpg)
stat_mode(mtcars$cyl)
}
\seealso{
\code{\link[base]{mean}} \code{\link[stats]{median}}
}

