% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{stat_range}
\alias{stat_range}
\title{Range}
\usage{
stat_range(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
Range of \code{x}
}
\description{
Compute the range of a numeric vector
}
\examples{
stat_range(mtcars$mpg)
}
\seealso{
\code{\link[base]{range}}
}
