% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-dist.R
\name{t_per}
\alias{t_per}
\title{Visualize T Distribution Percentile}
\usage{
t_per(probs = 0.95, df = 4, type = c("lower", "upper", "both"))
}
\arguments{
\item{probs}{a probability value}

\item{df}{degrees of freedom}

\item{type}{lower tail, upper tail or both}
}
\value{
percentile for the \code{probs} based on \code{df} and
\code{type}
}
\description{
Calculate and visualize quantiles out of given probability
}
\examples{
t_per(probs = 0.95, df = 4, type = 'lower')
t_per(probs = 0.35, df = 4, type = 'upper')
t_per(probs = 0.69, df = 7, type = 'both')
}
\seealso{
\code{\link{t_plot}} \code{\link{t_prob}} \code{\link[stats]{TDist}}
}

