% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-dist.R
\name{t_prob}
\alias{t_prob}
\title{T Distribution Probability}
\usage{
t_prob(perc, df, type = c("lower", "upper", "interval", "both"))
}
\arguments{
\item{perc}{t statistic}

\item{df}{degrees of freedom}

\item{type}{lower tail, upper tail, interval or both}
}
\value{
probability value for the \code{perc} based on \code{df} and
\code{type}
}
\description{
Visualize cumulative probabilities based on t statistics
}
\examples{
t_prob(2.045, 7, 'lower')
t_prob(0.945, 7, 'upper')
t_prob(1.445, 7, 'interval')
t_prob(1.6, 7, 'both')
}
\seealso{
\code{\link{t_plot}} \code{\link{t_per}} \code{\link[stats]{TDist}}
}

