% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult-table.R
\name{tway_tables}
\alias{tway_tables}
\title{Multiple Two Way Tables}
\usage{
tway_tables(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
\code{tway_tables} creates multiple two way tables by creating
a cross table for each unique pair of categorical variables in a data frame.
}
\details{
\code{tway_tables} is a extension of the \code{cross_table} 
function. It creates a two way table for each unique pair of categorical 
variables in the dataframe.
}
\examples{
mt <- mtcars
mt[, c(2, 8:11)] <- lapply(mt[, c(2, 8:11)], factor)
tway_tables(mt)
}
\seealso{
\code{link{oway_tables}} \code{link{freq_table}} 
\code{link{cross_table}}
}

