% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-dist.R
\name{chi_prob}
\alias{chi_prob}
\title{Visualize Chi Square Distribution Probabilities}
\usage{
chi_prob(perc, df, type = c("lower", "upper"))
}
\arguments{
\item{perc}{percentile}

\item{df}{degrees of freedom}

\item{type}{lower tail or upper tail}
}
\value{
probability value for \code{perc} based on \code{df} and \code{type}
}
\description{
Calculate and visualize probability from a given quantile.
}
\examples{
chi_prob(13.58, 11, 'lower')
chi_prob(15.72, 13, 'upper')
}
\seealso{
\code{\link{chi_plot}} \code{\link{chi_per}}
\code{\link[stats]{Chisquare}}
}

