% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-table.R
\name{cross_table}
\alias{cross_table}
\alias{cross_table.default}
\title{Two Way Tables}
\usage{
cross_table(var1, var2)

\method{cross_table}{default}(var1, var2)
}
\arguments{
\item{var1}{First categorical variable}

\item{var2}{Second categorical variable}
}
\value{
\code{cross_table} returns an object of class \code{"cross_table"}.
An object of class \code{"cross_table"} is a list containing at least the
following components:

\item{obs}{obs number of observations}
\item{var2_levels}{levels of the second categorical variable}
\item{var1_levels}{levels of the first categorical variable}
\item{varnames}{names of the variables}
\item{twowaytable}{table of the variables}
\item{percent_table}{table of percentages}
\item{row_percent}{table of row percentages}
\item{col_percent}{table of column percentages}
\item{column_totals}{total of columns}
\item{percent_column}{total of columns as a percentage}
}
\description{
\code{cross_table} creates two way tables of categorical
variables. The tables created can be visualized as barplots and mosaicplots.
}
\details{
cross_table is a generic function
}
\examples{
k <- cross_table(mtcars$cyl, mtcars$am)

# bar plots
plot(k)
plot(k, beside = TRUE)
plot(k, proportional = TRUE)

# mosaic plots
mosaicplot(k)
}

