% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-auto-summary.R
\name{ds_auto_summary_stats}
\alias{ds_auto_summary_stats}
\title{Summary statistics}
\usage{
ds_auto_summary_stats(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Generate summary statistics for all continuous variables in data.
}
\examples{
ds_auto_summary_stats(mtcarz)
ds_auto_summary_stats(mtcarz, disp, hp)

}
