% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_css}
\alias{ds_css}
\title{Corrected Sum of Squares}
\usage{
ds_css(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose mode is to be computed}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Corrected sum of squares of \code{x}
}
\description{
Compute the corrected sum of squares
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
ds_css(mtcars$mpg)
}
\references{
\href{http://www.itl.nist.gov/div898/handbook/prc/section4/prc421.htm}{NIST/SEMATECH e-Handbook of Statistical Methods}
}
