% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_cvar}
\alias{ds_cvar}
\title{Coefficient of Variation}
\usage{
ds_cvar(x, data = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{data}{a \code{data.frame} or \code{tibble}}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Compute the coefficient of variation
}
\examples{
ds_cvar(mtcars$mpg)
ds_cvar(mpg, mtcars)
}
