% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_cvar}
\alias{ds_cvar}
\title{Coefficient of Variation}
\usage{
ds_cvar(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose mode is to be computed}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Compute the coefficient of variation
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
ds_cvar(mtcars$mpg)
}
