% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_cvar}
\alias{ds_cvar}
\title{Coefficient of Variation}
\usage{
ds_cvar(data, x = NULL)
}
\arguments{
\item{data}{A numeric vector or \code{data.frame}.}

\item{x}{Column in \code{data}.}
}
\description{
Compute the coefficient of variation
}
\examples{
# vector
ds_cvar(mtcars$mpg)

# data.frame
ds_cvar(mtcars, mpg)

}
