% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_gmean}
\alias{ds_gmean}
\title{Geometric Mean}
\usage{
ds_gmean(x, data = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Computes the geometric mean
}
\examples{
ds_gmean(mtcars$mpg)
ds_gmean(mpg, mtcars)
}
\seealso{
\code{\link{ds_hmean}} \code{\link[base]{mean}}
}
