% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_hmean}
\alias{ds_hmean}
\alias{hmean}
\title{Harmonic Mean}
\usage{
ds_hmean(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector containing the values whose harmonic mean is to be
computed}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{...}{further arguments passed to or from other methods
#' @details Any NA values are stripped from \code{x} before computation
takes place.}
}
\value{
Returns the harmonic mean of \code{x}
}
\description{
Compute the harmonic mean
}
\section{Deprecated function}{

\code{hmean()} has been deprecated. Instead use \code{ds_hmean()}.
}

\examples{
ds_hmean(mtcars$mpg)
}
\seealso{
\code{\link{ds_gmean}} \code{\link[base]{mean}}
}
