% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_measures_location}
\alias{ds_measures_location}
\title{Measures of location}
\usage{
ds_measures_location(data, column, trim = 0.05)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{column}{Column in \code{data}.}

\item{trim}{The fraction of values to be trimmed before computing
the mean.}
}
\description{
Returns the measures of location such as mean, median & mode.
}
\examples{
ds_measures_location(mtcarz, mpg)

}
