% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_measures_symmetry}
\alias{ds_measures_symmetry}
\title{Measures of symmetry}
\usage{
ds_measures_symmetry(data, column)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{column}{Column in \code{data}.}
}
\description{
Returns the measures of symmetry such as skewness and kurtosis.
}
\examples{
ds_measures_symmetry(mtcarz, mpg)

}
