% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_measures_variation}
\alias{ds_measures_variation}
\title{Measures of variation}
\usage{
ds_measures_variation(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Returns the measures of location such as range, variance and standard
  deviation.
}
\examples{
ds_measures_variation(mtcarz)
ds_measures_variation(mtcarz, mpg)
ds_measures_variation(mtcarz, mpg, disp)

}
