% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_mode}
\alias{ds_mode}
\title{Mode}
\usage{
ds_mode(data, x = NULL)
}
\arguments{
\item{data}{A numeric vector or \code{data.frame}.}

\item{x}{Column in \code{data}.}
}
\value{
Mode of \code{x}
}
\description{
Compute the sample mode
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
# vector
ds_mode(mtcars$mpg)

# data.frame
ds_mode(mtcars, mpg)

}
\seealso{
\code{\link[base]{mean}} \code{\link[stats]{median}}
}
