% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-multistats.R
\name{ds_multi_stats}
\alias{ds_multi_stats}
\alias{multistats}
\title{Multiple Variable Statistics}
\usage{
ds_multi_stats(x, ...)
}
\arguments{
\item{x}{a tibble/data.frame}

\item{...}{columns in \code{x}}
}
\value{
a tibble
}
\description{
Summary/descriptive statistics for multiple variables
}
\section{Deprecated Function}{

\code{multistats()} has been deprecated. Instead use \code{ds_multi_stats()}
}

\examples{
ds_multi_stats(mtcarz, mpg, disp, hp)
}
