% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_percentiles}
\alias{ds_percentiles}
\title{Percentiles}
\usage{
ds_percentiles(data, column)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Returns the percentiles
}
\examples{
ds_percentiles(mtcarz)
ds_percentiles(mtcarz, mpg)
ds_percentiles(mtcarz, mpg, disp)

}
