% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_bar_grouped}
\alias{ds_plot_bar_grouped}
\title{Generate grouped bar plots}
\usage{
ds_plot_bar_grouped(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Creates grouped bar plots if the data has categorical variables.
}
\examples{
ds_plot_bar_grouped(mtcarz)
ds_plot_bar_grouped(mtcarz, cyl, gear)

}
