% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_bar_stacked}
\alias{ds_plot_bar_stacked}
\title{Generate stacked bar plots}
\usage{
ds_plot_bar_stacked(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Creates stacked bar plots if the data has categorical variables.
}
\examples{
ds_plot_bar_stacked(mtcarz)
ds_plot_bar_stacked(mtcarz, cyl, gear)

}
