% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_box_group}
\alias{ds_plot_box_group}
\title{Compare distributions}
\usage{
ds_plot_box_group(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Creates box plots if the data has both categorical & continuous variables.
}
\examples{
mt <- dplyr::select(mtcarz, cyl, disp, mpg)
ds_plot_box_group(mt)
ds_plot_box_group(mtcarz, cyl, gear, mpg)

}
