% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plots.R
\name{ds_plot_density}
\alias{ds_plot_density}
\title{Generate density plots}
\usage{
ds_plot_density(data, ..., color = "blue", print_plot = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}

\item{color}{Color of the plot.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Creates density plots if the data has continuous variables.
}
\examples{
# plot single variable
ds_plot_density(mtcarz, mpg)

# plot multiple variables
ds_plot_density(mtcarz, mpg, disp, hp)

# plot all variables
ds_plot_density(mtcarz)

}
