% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_rindex}
\alias{ds_rindex}
\title{Index Values}
\usage{
ds_rindex(data, values)
}
\arguments{
\item{data}{a numeric vector}

\item{values}{a numeric vector containing the values whose index is returned}
}
\value{
Index of the \code{values} in \code{data}. In case, \code{data} does
not contain \code{index}, \code{NULL} is returned.
}
\description{
Returns index of values.
}
\details{
Any NA values are stripped from \code{data} and \code{values} before
computation takes place.
}
\examples{
ds_rindex(mtcars$mpg, 21)
ds_rindex(mtcars$mpg, 22)
}
