% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-screener.R
\name{ds_screener}
\alias{ds_screener}
\alias{plot.ds_screener}
\title{Screen data}
\usage{
ds_screener(y)

\method{plot}{ds_screener}(x, ...)
}
\arguments{
\item{y}{A \code{tibble} or a \code{data.frame}.}

\item{x}{An object of class \code{ds_screener}.}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
\code{ds_screener()} returns an object of class \code{"ds_screener"}.
An object of class \code{"ds_screener"} is a list containing the
following components:

\item{Rows}{Number of rows in the data frame.}
\item{Columns}{Number of columns in the data frame.}
\item{Variables}{Names of the variables in the data frame.}
\item{Types}{Class of the variables in the data frame.}
\item{Count}{Length of the variables in the data frame.}
\item{nlevels}{Number of levels of a factor variable.}
\item{levels}{Levels of factor variables in the data frame.}
\item{Missing}{Number of missing observations in each variable.}
\item{MissingPer}{Percent of missing observations in each variable.}
\item{MissingTotal}{Total number of missing observations in the data frame.}
\item{MissingTotPer}{Total percent of missing observations in the data frame.}
\item{MissingRows}{Total number of rows with missing observations in the
data frame.}
\item{MissingCols}{Total number of columns with missing observations in the
data frame.}
}
\description{
Screen data  and return details such as variable names, class, levels and
missing values. \code{plot.ds_screener()} creates bar plots to visualize %
of missing observations for each variable in a data set.
}
\examples{
# screen data
ds_screener(mtcarz)

}
