% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-describe.R
\name{ds_skewness}
\alias{ds_skewness}
\title{Skewness}
\usage{
ds_skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose skewness is to be computed}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
Skewness of \code{x}
}
\description{
Compute the skewness of a probability distribution.
}
\details{
Any NA values are stripped from \code{x} before computation
takes place.
}
\examples{
ds_skewness(mtcars$mpg)
}
\references{
Sheskin, D.J. (2000) Handbook of Parametric and Nonparametric Statistical Procedures, Second Edition. Boca Raton, Florida: Chapman & Hall/CRC.
}
\seealso{
\code{kurtosis}
}
