% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-summary-stats.R
\name{ds_summary_stats}
\alias{ds_summary_stats}
\title{Descriptive statistics}
\usage{
ds_summary_stats(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{...}{Column(s) in \code{data}.}
}
\description{
Range of descriptive statistics for continuous data.
}
\examples{
# single variable
ds_summary_stats(mtcarz, mpg)

# multiple variables
ds_summary_stats(mtcarz, mpg, disp, hp)

# all variables
ds_summary_stats(mtcarz)

}
\seealso{
\code{\link[base]{summary}}
\code{\link{ds_freq_table}} \code{\link{ds_cross_table}}
}
