% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{gmean}
\alias{gmean}
\title{Geometric Mean}
\usage{
gmean(x, ...)
}
\arguments{
\item{x}{a numeric vector containing the values whose geometric mean is to be
computed}

\item{...}{further arguments passed to or from other methods}
}
\value{
Returns the geometric mean of x
}
\description{
Compute the geometric mean
}
\examples{
gmean(mtcars$mpg)
}
\seealso{
\code{\link{hmean}}
}

