% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{hmean}
\alias{hmean}
\title{Harmonic Mean}
\usage{
hmean(x, ...)
}
\arguments{
\item{x}{a numeric vector containing the values whose harmonic mean is to be
computed}

\item{...}{further arguments passed to or from other methods
#' @details Any NA values are stripped from \code{x} before computation
takes place.}
}
\value{
Returns the harmonic mean of \code{x}
}
\description{
Compute the harmonic mean
}
\examples{
hmean(mtcars$mpg)
}
\seealso{
\code{\link{gmean}} \code{\link[base]{mean}}
}
