% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-dist.R
\name{t_plot}
\alias{t_plot}
\title{T Distribution Plot}
\usage{
t_plot(df = 3)
}
\arguments{
\item{df}{degrees of freedom}
}
\description{
Visualize how degrees of freedom affect the shape of t
distribution
}
\examples{
t_plot()
t_plot(6)
t_plot(df = 8)
}
\seealso{
\code{\link{t_per}} \code{\link{t_prob}} \code{\link[stats]{TDist}}
}

